fix(serializer): remove unnecessary boxing conversions in unsigned types
- Removed redundant byteValue() call in cuint8_t serializer
- Removed redundant byteValue() call in cuchar serializer
- Removed redundant shortValue() call in cushort serializer
- Directly pass wrapper values to write methods for better performance- Maintain same serialization behavior while improving code clarity

+++ src/main/java/org/fz/nettyx/serializer/struct/basic/c/stdint/unsigned/cuint8_t.java
@@ -30,7 +30,7 @@ public class cuint8_t extends cbasic<Short> {
 
     @Override
     protected ByteBuf toByteBuf(Short value) {
-        return Unpooled.buffer(size).writeByte(value.byteValue());
+        return Unpooled.buffer(size).writeByte(value);
     }
 
     @Override
